checkout post mapping

+++ Fashion-backend/src/main/java/com/Fashionbackend/Repo/OrderRepo.java
@@ -3,5 +3,5 @@ package com.Fashionbackend.Repo;
 import com.Fashionbackend.Entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepo extends JpaRepository<Order, Long> {
+public interface OrderRepo extends JpaRepository<Order, Integer> {
 }
